Create helpers, move PartStaff routines to mixin
Just trying to keep down the size of any one module as much as possible, also makes it easier to diagnose and test where things are happening/going wrong
