+++ music21/musicxml/__init__.py
@@ -9,11 +9,15 @@
 # Copyright:    Copyright © 2010-2016 Michael Scott Cuthbert and the music21 Project
 # License:      BSD, see license.txt
 # ------------------------------------------------------------------------------
-__all__ = ['archiveTools', 'lilypondTestSuite', 'm21ToXml', 'xmlObjects', 'xmlToM21']
+__all__ = [
+    'archiveTools', 'lilypondTestSuite', 'm21ToXml',
+    'partStaffExporter', 'xmlObjects', 'xmlToM21',
+]
 
 from music21.musicxml import archiveTools
 from music21.musicxml import lilypondTestSuite
 from music21.musicxml import m21ToXml
+from music21.musicxml import partStaffExporter
 from music21.musicxml import xmlObjects
 from music21.musicxml import xmlToM21
